Avoid using nonstandard assembly syntax for NEON
While the original form of the example can't be assembled with
Clang, the suggested replacement isn't ideal either.

The original form of the example used a nonstandard way of
referencing vector elements, by writing `V3.4H[0]`, while the
standard form is to drop the total number of elements, writing
only `V3.H[0]` - this is supported by both GNU assembler and Clang.

The earlier replacement syntax, with instructions like `MUL.4H`,
where the vector layout is specified as part of the instruction
instead of as part of the register, is entirely nonstandard. (It
is similar to how NEON instructions were written for AArch32
though.) This syntax is accepted by Clang, and is produced by
LLVM tools when disassembling Mach-O file, but should not be
recommended in new written code.

Thus revert these instructions back to their original form
and only apply the element specifier syntax fix, changing
`V3.4H[0]` into `V3.H[0]`.